chia component, tạo login register cơ bản

+++ database/migrations/2024_11_21_091214_create_notifications_table.php
@@ -8,10 +8,10 @@ return new class {
         if(!Capsule::schema()->hasTable('notifications')){
             Capsule::schema()->create('notifications', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('user_id');
+                $table->unsignedBigInteger('user_id');
                 $table->text('title');
                 $table->text('message');
-                $table->unsignedInteger('office_id');
+                $table->unsignedBigInteger('office_id');
                 $table->timestamps();
 
           $table->foreign('user_id')
